+++ controllers/helpers.go
@@ -14,14 +14,12 @@ import (
 	"time"
 
 	"github.com/go-logr/logr"
-	lagoonv1alpha1 "github.com/uselagoon/remote-controller/api/v1alpha1"
+	lagoonv1alpha1 "github.com/uselagoon/remote-controller/apis/lagoon-old/v1alpha1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
-	// LabelAppManaged for discovery.
-	LabelAppManaged = "lagoon.amazee.io/managed-by"
 	// DefaultNamespacePattern is what is used when one is not provided.
 	DefaultNamespacePattern = "${project}-${environment}"
 )
