+++ handlers/helpers.go
@@ -11,13 +11,11 @@ import (
 	"strings"
 	"time"
 
-	lagoonv1alpha1 "github.com/uselagoon/remote-controller/api/v1alpha1"
+	lagoonv1alpha1 "github.com/uselagoon/remote-controller/apis/lagoon-old/v1alpha1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 )
 
 const (
-	// LabelAppManaged for discovery.
-	LabelAppManaged = "lagoon.amazee.io/managed-by"
 	// DefaultNamespacePattern is what is used when one is not provided.
 	DefaultNamespacePattern = "${project}-${environment}"
 )
